backend/requirements.txt
Flask==2.2.5
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
